==> docs/notebooks/03_transcriptome_reconstruction.ipynb <==
# Transcriptome Reconstruction
This tutorial processes the example data set, based on PacBio Isoseq samples of hematopoetic cells from ENCODE. This dataset contains only a subset of genomic regions, allowing for fast processing of the demonstration tutorials. 
All required data files to run the tutorials can be obtained here: ([download link](https://nc.molgen.mpg.de/cloud/index.php/s/zYe7g6qnyxGDxRd)). 

To run this example, you will you will need the following **input files**:

* sample description file 'encode_samples.tsv'
* six .bam alignment files for the six samples
* six corresponding .bam.bai indices
* referernce annotation file gencode.v42.chr_patch_hapl_scaff.annotation_sorted_chr8.gff3.gz
* corresponding .gff3.gz.tbi index file
* genomic reference file GRCh38.p13.genome_chr8.fa
* cooresponding .fai index file

All files are assumed to be stored in a subfolder called 'demonstration_dataset'

In this tutorial, we import the reference annotation, specify the alignment files for the samples, and integrate the data into a common data structure. During this step, the transcriptome is reconstructed, and quality control metrics are computed. 


# preperation: import the libraries
from  isotools import Transcriptome
from isotools import __version__ as isotools_version
import pandas as pd
import matplotlib.pyplot as plt
import logging
# set up logging
logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)
logger=logging.getLogger('isotools')
logger.info(f'This is isootools version {isotools_version}')


path='demonstration_dataset'


## Import of reference annotation
The first step is to import the reference annotation from a gff or gtf file with the [Transcriptome.from_reference](../isotoolsAPI.html?highlight=from_reference#isotools.Transcriptome.from_reference) class method. The input file should be sorted and indexed with tabix. 

annotation_fn=f'{path}/gencode.v42.chr_patch_hapl_scaff.annotation_sorted_chr8.gff3.gz'
#create the IsoTools transcriptome object from the reference annotation
isoseq=Transcriptome.from_reference(annotation_fn)


## Import of sequencing data
Next a sample table containing the sample names as well as corresponding file names and group assignment is imported.
This information is used in the next step, to reconstruct the transcripts for each sample and integrate them with the transcriptome object, using the [add_sample_from_bam](../isotoolsAPI.html?highlight=add_sample_from_bam#isotools.Transcriptome.add_sample_from_bam) function.
When all samples are added, quality control metrics are calculated by calling [Transcriptome.add_qc_metrics](../isotoolsAPI.html?highlight=add_qc_metrics#isotools.Transcriptome.add_qc_metrics).
Last, the object with all data is stored on disk in a pickle file for later use with the [save](../isotoolsAPI.html?highlight=save#isotools.Transcriptome.save) method.
In another session, the object can be restored with the classmethod [Transcriptome.load](../isotoolsAPI.html?highlight=load#isotools.Transcriptome.load).

sample_fn=f'{path}/encode_samples.tsv'
genome_fn=f'{path}/GRCh38.p13.genome_chr8.fa'

samples=pd.read_csv(sample_fn, sep='\t')
samples.file_name=path+'/'+samples.file_name
samples

# integrate the samples
for i,row in samples.iterrows():
    # this step takes about 5-30 seconds per sample
     isoseq.add_sample_from_bam(row.file_name, sample_name=row.sample_name, group=row.group)
# the sample table of the transcriptome object contains the number of imported reads
isoseq.sample_table

In the next step, we compute several qc metrics for the transcripts: 

* downstream A content, 
* direct repeat length at junctions, 
* noncanonical splicing, 
* potential fragments

This information is stored with the genes objects, and can be accessed by downstream analysis. 

# compute qc metrics
isoseq.add_qc_metrics(genome_fn)
# add ORF predictions
isoseq.add_orf_prediction(genome_fn)


## Transcriptome Export
The identified transcrips can be exported in pickle format, GTF format, and as a table.
* Pickle format is an interlal python format, used to save the entire transcriptome data, so it can be restored in an IsoTools session without reimporting the alignment files. Export to pickle is done with the [save](../isotoolsAPI.html?highlight=save#isotools.Transcriptome.save) methods
* GTF (General Feature Format) is a file format commonly used in bioinformatics to represent annotated genomic features, such as gene models, exons, and introns. Exporting to this format is done with the [write_gtf](../isotoolsAPI.html?highlight=write_gtf#isotools.Transcriptome.write_gtf) method, and facilitates the use of the reconstructed transcriptome in several external tools. 
* The transcripts can also be exported as a table, containing coverage information, as well as additional features such as qc metrics, specified with the "extra_columns" parameter. The use of the [transcript_table](../isotoolsAPI.html?highlight=transcript_table#isotools.Transcriptome.transcript_table) methods is explained in the API documentation. 


To select the transcripts to be exported to GTF and the table, we used the isotools filtering query syntax, which is explained in detail in a [seperate tutorial](06_filtering.html#Filtering-tags-and-queries). 
The gtf can be compressed by adding a '.gz' to the filename and setting  gzip=True. 
The table can be set to contain samplewise and/or groupwise coverage information, by setting the sample and the groups parameters. Here we sum the read counts per group. 


# export the transcriptome object for later use. 
isoseq.save(f'{path}/PacBio_isotools.pkl')
# to load the data in the next session, use 
# isoseq=Transcriptome.load('PacBio_isotools_substantial_isotools.pkl')

# export gtf:
query_string = 'SUBSTANTIAL and not (NOVEL_TRANSCRIPT and UNSPLICED)'
isoseq.write_gtf(f'{path}/demonstration_dataset_substantial_transcripts.gtf', 
                 source='isoseq', min_coverage=5,  
                 gzip=False, 
                 query=query_string)



# export transcript table with the same filter criteria:
transcript_tab=isoseq.transcript_table( groups=isoseq.groups(),tpm=True,coverage=True, 
                                       min_coverage=5, progress_bar=True, 
                                       query=query_string)
# write to csv file
transcript_tab.to_csv(f'{path}/demonstration_dataset_substantial_transcripts.csv', 
                      index=False, sep='\t')
#show the first lines
transcript_tab.head()
